Split out oneOffContributions and generic fetch-related functions
